refactor: 오토컴플리트api 리무브 및 그 대신 useallmaps씀
